+++ R/ncdfCF-package.R
@@ -59,11 +59,13 @@
 #'
 #' * `show()`: Print information about the data set to the console. This will
 #' print some information on the resource, as well as all identified data
-#' variables, axes, and the global attributes.
+#' variables and the global attributes.
 #' * `name`: The name of the data set.
 #' * `conventions`: The conventions that this data set complies with. There
 #' could be multiple conventions listed in a single character string. CF
 #' Metadata Conventions use a "CF-1.*" label.
+#' * `var_names()`, `axis_names`: Return a character vector with the names of the
+#' objects.
 #' * `variables()`, `axes()`, `attributes()`: Return a `list` or `data.frame` with
 #' all objects of the specific type found in the data set.
 #' * `find_by_name()`: Find a named object in the data set. This can be a data